Unify sound settings screens
